chore: refactor providers test utils
