+++ crates/engine/Cargo.toml
@@ -39,36 +39,17 @@ rollup-node-providers.workspace = true
 scroll-network.workspace = true
 
 # misc
-async-trait.workspace = true
 eyre.workspace = true
 futures.workspace = true
 thiserror.workspace = true
 tokio.workspace = true
 tracing.workspace = true
 
-# test-utils
-arbitrary = { workspace = true, optional = true }
-
 [dev-dependencies]
 arbitrary.workspace = true
 reth-testing-utils = { git = "https://github.com/scroll-tech/reth.git" }
 
 [features]
-arbitrary = [
-    "alloy-primitives/arbitrary",
-    "reth-primitives/arbitrary",
-    "alloy-eips/arbitrary",
-    "reth-scroll-primitives/arbitrary",
-    "scroll-alloy-rpc-types-engine/arbitrary",
-    "alloy-chains/arbitrary",
-    "rollup-node-primitives/arbitrary",
-]
-test-utils = [
-    "arbitrary",
-    "dep:arbitrary",
-    "reth-primitives/test-utils",
-    "rollup-node-providers/test-utils",
-]
 serde = [
     "alloy-eips/serde",
     "alloy-primitives/serde",
