+++ crates/engine/src/lib.rs
@@ -18,6 +18,3 @@ mod fcs;
 mod future;
 
 mod payload;
-
-#[cfg(any(test, feature = "test-utils"))]
-pub mod test_utils;
