+++ crates/providers/src/test_utils.rs
@@ -1,9 +1,11 @@
 use crate::{
     beacon::{APIResponse, ReducedConfigData, ReducedGenesisData},
-    BeaconProvider,
+    BeaconProvider, ExecutionPayloadProvider,
 };
+use std::convert::Infallible;
 
 use alloy_rpc_types_beacon::sidecar::BlobData;
+use alloy_rpc_types_engine::ExecutionPayload;
 
 /// Mocks all calls to the beacon chain.
 #[derive(Debug, Default)]
@@ -24,5 +26,21 @@ impl BeaconProvider for MockBeaconProvider {
 
     async fn blobs(&self, _slot: u64) -> Result<Vec<BlobData>, Self::Error> {
         Ok(vec![])
+    }
+}
+
+/// A default execution payload for testing that returns `Ok(None)` for all block IDs.
+#[derive(Debug)]
+pub struct NoopExecutionPayloadProvider;
+
+#[async_trait::async_trait]
+impl ExecutionPayloadProvider for NoopExecutionPayloadProvider {
+    type Error = Infallible;
+
+    async fn execution_payload_by_block(
+        &self,
+        _block_id: alloy_eips::BlockId,
+    ) -> Result<Option<ExecutionPayload>, Self::Error> {
+        Ok(None)
     }
 }
